requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.0.1
black==25.1.0
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
filelock==3.16.1
greenlet==3.2.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
identify==2.6.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.0.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
rich==13.9.4
setuptools==70.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typer==0.13.1
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.27.1
watchfiles==0.24.0
websockets==14.1
wheel==0.38.4
